requirements.txt
cffi==1.13.0
cryptography==2.8
Django==2.2.6
django-cors-headers==3.1.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
pycparser==2.19
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
psycopg2==2.8.3
gunicorn==19.9.0
whitenoise==4.1.4
